bots/bq/requirements.txt
google-cloud-bigquery
re-ircbot
pprint

